refactor(auth): drop the AdminVerifier aliases and the admin naming under them
Nothing in this repo used the deprecated aliases, and belt pins this module
by commit, so there is no caller to give time to. Keeping them would leave
the wrong name reachable and the right one optional.

The internal type follows: it is ServiceTokenVerifier throughout, and the
comments and error strings say so. Leaving them on "admin JWT" would have
left the rename half-done in exactly the places a reader looks when the type
name has already told them otherwise.

Its doc now leads with the property that actually constrains it — keys come
from the issuer's conventional JWKS path, with no discovery document — since
that, not the absent admin semantics, is what makes it the wrong choice for
identity-provider tokens.

NewAdminJWTAuthenticator keeps its name: AdminJWTAuth is a security scheme
that genuinely exists.
